:wrench: modal state triggering successful
